Bitrunning: Combat domain [READY] (#84196)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Finally dusts off this project to make a deathmatch style bitrunning
map.
Don't be too intimidated by the file diff, lots of code organization +
resized a large map.

Changes:

1. Reuses the gateway beach map as a combat zone (99% of the file diff)
(maptainers: i just added spawners and areas)
2. Alters how bitrunning handles spawning: Custom spawns are now
available, which can be anything

Misc organization:

- Splits netpod.dm into separate files.
- Fixes some wording in vdom map documentation.
- Organizes vdom variables a bit.
- Adds a permanent hololadder spawn.

How bitrunning deathmatch works:

- Temporary spawners are offered to both ghosts and bitrunners.
- Runners spawn in like usual. Ghost can use the spawner menu.
- Ghosts work to prevent avatars from collecting side objectives or try
to cause mass brain damage.
- The domain completes after a number of deaths accrue. Any faction.
Blood for the blood god, etc.
- This map can be played solo. ANY deaths.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

I've been toying with the idea of a deathmatch style map for some time.
I liked syndicate assault, the spawners were intentionally left there,
and the possibility of player-controlled players made the experience
more tense and challenging.

This PR leans into this idea: The virtual world is dangerous. Players
get a chance to compete on both sides here. It offers a lot of variety
to bitrunning other than "run for box". It's also very lucrative if
ghosts join in.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Added a bitrunning deathmatch map: Island Brawl. Both ghosts and
runners get many more spawns than normal.
fix: Lowered the static vision time in domain load in.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

+++ code/modules/bitrunning/server/obj_generation.dm
@@ -15,9 +15,31 @@
 	new /obj/structure/closet/crate/secure/bitrunning/encrypted(chosen_turf)
 	return TRUE
 
+
+/// Attempts to spawn a lootbox
+/obj/machinery/quantum_server/proc/attempt_spawn_curiosity(list/possible_turfs)
+	if(!length(possible_turfs)) // Out of turfs to place a curiosity
+		return FALSE
+
+	if(generated_domain.secondary_loot_generated >= assoc_value_sum(generated_domain.secondary_loot)) // Out of curiosities to place
+		return FALSE
+
+	shuffle_inplace(possible_turfs)
+	var/turf/chosen_turf = validate_turf(pick(possible_turfs))
+
+	if(isnull(chosen_turf))
+		possible_turfs.Remove(chosen_turf)
+		chosen_turf = validate_turf(pick(possible_turfs))
+		if(isnull(chosen_turf))
+			CRASH("vdom: after two attempts, could not find a valid turf for curiosity")
+
+	new /obj/item/storage/lockbox/bitrunning/encrypted(chosen_turf)
+	return chosen_turf
+
+
 /// Generates a new avatar for the bitrunner.
-/obj/machinery/quantum_server/proc/generate_avatar(obj/structure/hololadder/wayout, datum/outfit/netsuit)
-	var/mob/living/carbon/human/avatar = new(wayout.loc)
+/obj/machinery/quantum_server/proc/generate_avatar(turf/destination, datum/outfit/netsuit)
+	var/mob/living/carbon/human/avatar = new(destination)
 
 	var/outfit_path = generated_domain.forced_outfit || netsuit
 	var/datum/outfit/to_wear = new outfit_path()
@@ -41,8 +63,9 @@
 	if(istype(hat))
 		hat.set_armor(/datum/armor/none)
 
-	for(var/obj/thing in avatar.held_items)
-		qdel(thing)
+	if(!generated_domain.forced_outfit)
+		for(var/obj/thing in avatar.held_items)
+			qdel(thing)
 
 	var/obj/item/storage/backpack/bag = avatar.back
 	if(istype(bag))
@@ -61,32 +84,16 @@
 
 		SSid_access.apply_trim_to_card(outfit_id, /datum/id_trim/bit_avatar)
 
+	avatar.AddComponent( \
+		/datum/component/simple_bodycam, \
+		camera_name = "bitrunner bodycam", \
+		c_tag = "Avatar [avatar.real_name]", \
+		network = BITRUNNER_CAMERA_NET, \
+		emp_proof = TRUE, \
+	)
+
 	return avatar
 
-/// Generates a new hololadder for the bitrunner. Effectively a respawn attempt.
-/obj/machinery/quantum_server/proc/generate_hololadder()
-	if(!length(exit_turfs))
-		return
-
-	if(retries_spent >= length(exit_turfs))
-		return
-
-	var/turf/destination
-	for(var/turf/dest_turf in exit_turfs)
-		if(!locate(/obj/structure/hololadder) in dest_turf)
-			destination = dest_turf
-			break
-
-	if(isnull(destination))
-		return
-
-	var/obj/structure/hololadder/wayout = new(destination, src)
-	if(isnull(wayout))
-		return
-
-	retries_spent += 1
-
-	return wayout
 
 /// Loads in any mob segments of the map
 /obj/machinery/quantum_server/proc/load_mob_segments()
@@ -114,6 +121,7 @@
 		qdel(landmark)
 
 	return TRUE
+
 
 /// Scans over neo's contents for bitrunning tech disks. Loads the items or abilities onto the avatar.
 /obj/machinery/quantum_server/proc/stock_gear(mob/living/carbon/human/avatar, mob/living/carbon/human/neo, datum/lazy_template/virtual_domain/generated_domain)
